requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
